==> 02-ggplot.Rmd <==
# Plotting data with ggplot

In the literature, the most common way to visualize \(\epsilon\)^205^Tl is plotted against 1/Tl concentration (ppm), so this is why I made the merged sheet during my last stage of cleaning. Thanks to recent practice from class assignments, plotting the data to share with my collaborators was the easieast part.

```{r making-plot, eval = TRUE, echo = TRUE, warning = FALSE}

# load packages

library(tidyverse)
library(dplyr)
library(ggplot2)
library(knitr)

# remind RMarkdown that iso_el = merged_data.csv (had some trouble with this)

iso_el <- read.csv("merged_data.csv")

# Tl isotope and 1/Tl  plot (note to self: low Tl conc outlier cropped out in this version)

tl_plot <- iso_el %>%
  ggplot(aes(x = Tl_inv, y = mean, col = combined_names)) +
  geom_point(aes(shape = month)) +
  geom_errorbar(aes(ymin = mean - sd2, ymax = mean + sd2), width = .2)+
  theme_light() + 
  labs(x = "1/Tl (ppm)", y = "Epsilon205Tl", col = "Site", shape = "Month") +
  theme(legend.position = c(0.85,0.3)) +
  xlim(0, 20) +
  ggsave(filename = "tl_plot2.tiff",
         device = "tiff",
         width = 8,
         height = 8,
         units = "in",
         dpi = 300)

tl_plot

```

```{r tl-fig, echo = TRUE, fig.align = "center", fig.cap = "The plot.", out.width = '40%'}

# hopefully insert the plot... it's taking forever to load!

knitr::include_graphics("tl_plot2.tiff")

```

My next step is to figure out how to get \(\epsilon\)^205^Tl on my axis label correctly. :)
