==> 03-database.Rmd <==
# Creating a database in SQLite

Here, I present the steps I have taken to build my metal_sources database in SQLite. After my comprehensive exams are finally over, I will populate this database with data from all other Tl isotope studies that have been conducted on metal sources and environmental samples (thankfully there are very few of these). For now, here is what I did to create the database for our previous assignment.

***

```{r nice-fig, echo = FALSE, fig.align = "center", fig.cap = "Figure illustrating metal_sources database schema.", out.width = '80%'}

knitr::include_graphics("metal-sources_schema.png")

```


```{r install, eval = FALSE, echo = TRUE}
# install packages

install.packages("RSQLite")
install.packages("DBI")

# load DBI package

library(DBI)

#  establish database connection

metal_sources_db <- dbConnect(drv = RSQLite::SQLite(), "/Users/mollyblakowski/Documents/professional/PhD/research/metal_sources/data/metal_sources.db")

```

My database has 8 tables. The first tables I need to create are sites and projects because they don't have foreign keys. These two tables house data associated with where samples were collected and with what project they are associated - specifically, is the data from my project, a collaborator’s project, or a project from the literature.

```{r firsttables, eval = FALSE, echo = TRUE}

# create sites and projects tables

dbExecute(metal_sources_db, "CREATE TABLE sites (
site_id char(3) NOT NULL PRIMARY KEY,
site_name varchar(50),
latitude double,
longitude double
);")

dbExecute(metal_sources_db, "CREATE TABLE projects (
project_id char(3) NOT NULL PRIMARY KEY,
contact_first varchar(50),
contact_last varchar(50),
doi varchar(50)
);")

```

Next, I can create the collections table, which stores information about when sample bottles were deployed and collected from sites. The reason this goes next is that I can't make the samples table without it.

```{r collections, eval = FALSE, echo = TRUE}

# create collections table

dbExecute(metal_sources_db, "CREATE TABLE collections (
collection_id INTEGER PRIMARY KEY,
deploy_date text,
collect_date text,
period_id varchar(20),
site_id char(3),
FOREIGN KEY (site_id) REFERENCES sites(site_id)
);")

```

The samples table contains information on the sample mass and the dimensions of the sampler that are needed to calculate flux. It is connected to collections and sites via collection_id, as as well to projects via project_id.

```{r samples, eval = FALSE, echo = TRUE}

# create samples table

dbExecute(metal_sources_db, "CREATE TABLE samples (
sample_id INTEGER PRIMARY KEY,
collection_id varchar(5),
mass_g float,
inlet_area_cm2 float,
inlet_area_m2 float,
project_id  char(3),
FOREIGN KEY (collection_id) REFERENCES collections(collection_id),
FOREIGN KEY (project_id) REFERENCES projects(project_id)
);")

```

The remaining tables are here to store results from my work in the lab.

```{r results, eval = FALSE, echo = TRUE}

# create grain_size table

dbExecute(metal_sources_db, "CREATE TABLE grain_size (
grain_size_id INTEGER PRIMARY KEY AUTOINCREMENT,
sample_id varchar(5),
median_grain_size_um float,
mean_grain_size_um float,
min_size_um float,
max_size_um float,
FOREIGN KEY (sample_id) REFERENCES samples(sample_id)
);")

# create isotopes table

dbExecute(metal_sources_db, "CREATE TABLE isotopes (
isotopes_id INTEGER PRIMARY KEY AUTOINCREMENT,
sample_id varchar(5),
epsilon_tl float,
epsilon_tl_error float,
FOREIGN KEY (sample_id) REFERENCES samples(sample_id)
);")

# create minerals table

dbExecute(metal_sources_db, "CREATE TABLE minerals (
minerals_id INTEGER PRIMARY KEY AUTOINCREMENT,
sample_id varchar(5),
quartz_percent float,
calcite_percent float,
halite_percent float,
chlorite_percent float,
kaolinite_percent float,
aragonite_percent float,
andesine_percent float,
bytownite_percent float,
dolomite_percent float,
gypsum_percent float,
illite_percent float,
oligoclase_percent float,
plagioclase_percent float,
FOREIGN KEY (sample_id) REFERENCES samples(sample_id)
);")

# create elements table

dbExecute(metal_sources_db, "CREATE TABLE elements (
elements_id INTEGER PRIMARY KEY AUTOINCREMENT,
sample_id varchar(5),
as_ppm float,
cd_ppm float,
co_ppm float,
cr_ppm float,
cu_ppm float,
pb_ppm float,
tl_ppm float,
zr_ppm float,
FOREIGN KEY (sample_id) REFERENCES samples(sample_id)
);")

```

Wow, I made a lot of typos while setting this up. But I'm done! Well, I still need to import my data into these tables, and after I do that, I also need to create temp tables for the results tables as per the directions in our course materials.
